requirements.txt
asttokens==3.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
comm==0.2.3
cryptography==46.0.4
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-secret-manager==2.26.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
packaging==26.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.4
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzmq==27.1.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
stack-data==0.6.3
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.5.0
